#5 created invite and app  components, rename previous app to main component, install react-router-dom
