+++ package.json
@@ -24,6 +24,7 @@
     "react-google-maps": "^9.4.5",
     "react-google-maps-loader": "^4.2.3",
     "react-places-autocomplete": "^7.2.0",
+    "react-router-dom": "^4.3.1",
     "yelp-fusion": "^2.2.1"
   },
   "devDependencies": {
